Adicionados testes e alterada estrutura de diretórios
