+++ Dockerfile
@@ -1,5 +1,5 @@
 FROM golang:1.7
-COPY ibpt /go/ibpt
+COPY bin/ibpt /go/ibpt
 COPY ibpt.db /go/ibpt.db
 RUN chmod +x  /go/ibpt
 EXPOSE 8082
