--- glide.lock
@@ -1,10 +1,0 @@
-hash: b7181a4ffbb43a363ec3073b54573b5dd09fcd978db9c39d72ef50bf3b235ccf
-updated: 2017-01-12T22:21:38.56934148-02:00
-imports:
-- name: github.com/mattn/go-sqlite3
-  version: ca5e3819723d8eeaf170ad510e7da1d6d2e94a08
-- name: golang.org/x/net
-  version: 0e2717dc3cc05907dc23096ef3a9086ea93f567f
-  subpackages:
-  - context
-testImports: []
