--- glide.yaml
@@ -1,4 +1,0 @@
-package: github.com/compufour/IBPT
-import:
-- package: github.com/mattn/go-sqlite3
-  version: ~1.2.0
