+++ src/compufour/glide.yaml
@@ -1,0 +1,4 @@
+package: compufour
+import:
+- package: github.com/mattn/go-sqlite3
+  version: ~1.2.0
